+++ client/src/components/app.tsx
@@ -4,13 +4,13 @@ import Snackbar from "@material-ui/core/Snackbar";
 import { connect } from "react-redux";
 import { compose } from "recompose";
 import openSocket from "../actions/socket/openSocket";
-import Navbar from "./navbar";
-import MainComponent from "./main-component";
+import Navbar from "./dash-components/navbar";
+import MainComponent from "./main-components/main-component";
 import ArmComponent from "./arm-components/arm-component";
-import CameraComponent from "./camera-component";
-import TelemetryComponent from "./telemetry-component";
-import ScienceComponent from "./science-component";
-import DashComponent from "./dashboard";
+import CameraComponent from "./camera-components/camera-component";
+import TelemetryComponent from "./telemetry-components/telemetry-component";
+import ScienceComponent from "./science-components/science-component";
+import DashComponent from "./dash-components/dashboard";
 import makeRequest from "../utils/request/makeRequest";
 import { Route, HashRouter } from "react-router-dom";
 import { MuiThemeProvider, createMuiTheme } from "@material-ui/core/styles";
