+++ client/src/components/camera-components/camera-component.css
