+++ client/src/components/camera-components/camera-component.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import openSocket from "../actions/socket/openSocket";
+import openSocket from "../../actions/socket/openSocket";
 import PhotoUpdate1 from "./camera-feed-one-component";
 import PhotoUpdate2 from "./camera-feed-two-component";
 import PhotoUpdate3 from "./camera-feed-three-component";
 import PhotoUpdate4 from "./camera-feed-four-component";
 import "./camera-component.css";
-import DataPacket from "../types";
+import DataPacket from "../../types";
 
 // rename photoupdates
 type CameraProps = {
