+++ client/src/components/camera-components/camera-feed-four-component.tsx
