+++ client/src/components/camera-components/camera-feed-one-component.tsx
