+++ client/src/components/camera-components/camera-feed-three-component.tsx
