+++ client/src/components/camera-components/photo.css
