+++ client/src/components/dash-components/navbar.css
