+++ client/src/components/science-components/linear-rail-toggle-component.tsx
