+++ client/src/components/science-components/science-component.tsx
@@ -2,14 +2,14 @@ import { Typography, CssBaseline, withStyles, Theme } from "@material-ui/core";
 import * as React from "react";
 import { connect } from "react-redux";
 import { compose } from "recompose";
-import openSocket from "../actions/socket/openSocket";
+import openSocket from "../../actions/socket/openSocket";
 import CO2graph from "./scienceGraphCO2";
 import TVOCgraph from "./scienceGraphTVOC";
 import Button from '@material-ui/core/Button';
 import DrillToggle from './drill-toggle-component';
 import LinearRailToggle from './linear-rail-toggle-component';
 import DrillUpDownToggle from "./drill-updown-component";
-import DataPacket from "../types";
+import DataPacket from "../../types";
 
 type ScienceProps = {
     sensors: DataPacket;
