+++ client/src/components/telemetry-components/telemetry-component.tsx
@@ -2,9 +2,9 @@ import { Typography, CssBaseline, withStyles, Theme } from "@material-ui/core";
 import * as React from "react";
 import { connect } from "react-redux";
 import { compose } from "recompose";
-import openSocket from "../actions/socket/openSocket";
+import openSocket from "../../actions/socket/openSocket";
 import TelemetryTableComponent from "./telemetry-table-component";
-import DataPacket from "../types";
+import DataPacket from "../../types";
